# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "패션 이커머스 StyleHub 백엔드 API. 로그인이 필요한 API 는 POST /api/v1/users/login 으로 받은 SESSION 쿠키로 인증한다.",
    "title" : "StyleHub API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/api/v1/admin/stores" : {
      "get" : {
        "description" : "필요 역할: ADMIN",
        "operationId" : "getStores",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "SUSPENDED", "REJECTED" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StoreResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "스토어 목록 조회(상태 필터)",
        "tags" : [ "회원·스토어" ]
      }
    },
    "/api/v1/admin/stores/{storeId}" : {
      "get" : {
        "description" : "필요 역할: ADMIN",
        "operationId" : "getStore",
        "parameters" : [ {
          "in" : "path",
          "name" : "storeId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "스토어 상세 조회",
        "tags" : [ "회원·스토어" ]
      }
    },
    "/api/v1/admin/stores/{storeId}/status" : {
      "patch" : {
        "description" : "필요 역할: ADMIN",
        "operationId" : "updateStoreStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "storeId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StoreStatusUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "스토어 상태 변경(승인·거절·정지)",
        "tags" : [ "회원·스토어" ]
      }
    },
...
    "/api/v1/stores/me" : {
      "get" : {
        "description" : "필요 역할: STORE",
        "operationId" : "getMyStore",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "내 스토어 조회",
        "tags" : [ "회원·스토어" ]
      }
    },
...
    "/api/v1/users/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserLoginResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "로그인 — 응답의 SESSION 쿠키로 이후 API 를 인증한다",
        "tags" : [ "회원·스토어" ]
      }
    },
    "/api/v1/users/logout" : {
      "post" : {
        "operationId" : "logout",
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "로그아웃",
        "tags" : [ "회원·스토어" ]
      }
    },
...
    "/api/v1/users/oauth/{provider}" : {
      "get" : {
        "operationId" : "authorizationUrl",
        "parameters" : [ {
          "in" : "path",
          "name" : "provider",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "GOOGLE" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "소셜 로그인 인가 URL 발급",
        "tags" : [ "회원·스토어" ]
      }
    },
    "/api/v1/users/oauth/{provider}/callback" : {
      "get" : {
        "operationId" : "callback",
        "parameters" : [ {
          "in" : "path",
          "name" : "provider",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "GOOGLE" ]
          }
        }, {
          "in" : "query",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "state",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OAuthLoginResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "소셜 로그인 콜백",
        "tags" : [ "회원·스토어" ]
      }
    },
    "/api/v1/users/sign-up" : {
      "post" : {
        "operationId" : "signUp",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserSignUpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSignUpResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "일반 회원 가입",
        "tags" : [ "회원·스토어" ]
      }
    },
    "/api/v1/users/sign-up/store" : {
      "post" : {
        "operationId" : "signUpWithStore",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StoreSignUpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreSignUpResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "스토어 입점 신청(스토어 회원 가입)",
        "tags" : [ "회원·스토어" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "OAuthLoginResponse" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "newUser" : {
            "type" : "boolean"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "USER", "STORE", "ADMIN" ]
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "StoreResponse" : {
        "type" : "object",
        "properties" : {
          "approvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "SUSPENDED", "REJECTED" ]
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "StoreSignUpRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "name" : {
            "type" : "string",
            "maxLength" : 10,
            "minLength" : 2,
            "pattern" : "^[가-힣a-zA-Z0-9]+$"
          },
          "password" : {
            "type" : "string",
            "maxLength" : 15,
            "minLength" : 8,
            "pattern" : "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,15}$"
          },
          "storeDescription" : {
            "type" : "string",
            "maxLength" : 400,
            "minLength" : 0
          },
          "storeName" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 2
          }
        },
        "required" : [ "email", "name", "password", "storeDescription", "storeName" ]
      },
      "StoreSignUpResponse" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "USER", "STORE", "ADMIN" ]
          },
          "storeName" : {
            "type" : "string"
          },
          "storeStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "SUSPENDED", "REJECTED" ]
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "StoreStatusUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "SUSPENDED", "REJECTED" ]
          }
        },
        "required" : [ "status" ]
      },
...
      "UserLoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "email", "password" ]
      },
      "UserLoginResponse" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "USER", "STORE", "ADMIN" ]
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "UserSignUpRequest" : {
        "type" : "object",
        "properties" : {
          "birthDate" : {
            "type" : "string",
            "format" : "date"
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "name" : {
            "type" : "string",
            "maxLength" : 10,
            "minLength" : 2,
            "pattern" : "^[가-힣a-zA-Z0-9]+$"
          },
          "password" : {
            "type" : "string",
            "maxLength" : 15,
            "minLength" : 8,
            "pattern" : "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,15}$"
          }
        },
        "required" : [ "birthDate", "email", "name", "password" ]
      },
      "UserSignUpResponse" : {
        "type" : "object",
        "properties" : {
          "birthDate" : {
            "type" : "string",
            "format" : "date"
          },
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    },
    "securitySchemes" : {
      "session" : {
        "in" : "cookie",
        "name" : "SESSION",
        "type" : "apiKey"
      }
    }
  }
}